+++ app/components/admin/user_invitations_component.rb
@@ -10,7 +10,7 @@ class Admin::UserInvitationsComponent < ViewComponent::Base
     render(DescriptionListComponent.new) do |list|
       list.with_item(StatListItemComponent.new(label: "Available Invites", value: available_invites_value, key: "invitations.available_invites"))
       list.with_item(StatListItemComponent.new(label: "Created Invites", value: @stats.created_invites_count))
-      list.with_item(StatListItemComponent.new(label: "Invited Users", value: @stats.invited_users_count))
+      list.with_item(StatListItemComponent.new(label: "Invited Users", value: invited_users_value))
     end
   end
 
@@ -23,6 +23,11 @@ class Admin::UserInvitationsComponent < ViewComponent::Base
         helpers.link_to("Edit", "#", class: EDIT_CLASSES, data: edit_modal_data)
       ])
     end
+  end
+
+  def invited_users_value
+    links = @stats.invited_users.map { |invite| helpers.admin_user_link(invite.invited_user) }
+    helpers.safe_join(links, ", ")
   end
 
   def edit_modal_data